Removed helpers1 directory, moved helper test files to `helpers`
