+++ test/unit/public-sdk-apis/sendUniqueOutcome.ts
@@ -11,8 +11,8 @@ import MainHelper from "../../../src/helpers/MainHelper";
 import Log from "../../../src/libraries/Log";
 import Database from "../../../src/services/Database";
 import timemachine from "timemachine";
-import OutcomeTestHelper from '../helpers/OutcomeTestHelper';
 import { SentUniqueOutcome } from '../../../src/models/Outcomes';
+import OutcomeTestHelper from '../../support/tester/OutcomeTestHelper';
 
 const sinonSandbox: SinonSandbox = sinon.sandbox.create();
 const OUTCOME_NAME = "test_outcome";
